+++ core/src/console/connect_to_director.cc
@@ -28,6 +28,9 @@
 #include "lib/global_resource.h"
 #include "lib/bstringlist.h"
 #include "lib/bsock_tcp.h"
+#include "lib/hello.h"
+#include "lib/version.h"
+#include "lib/default_console.h"
 
 namespace console {
 BareosSocket* ConnectToDirector(JobControlRecord& jcr,
@@ -46,52 +49,34 @@ BareosSocket* ConnectToDirector(JobControlRecord& jcr,
   jcr.dir_bsock = UA_sock;
 
   const char* name;
-  s_password* password = NULL;
 
   TlsResource* local_tls_resource;
   if (console_resource) {
     name = console_resource->resource_name_;
     ASSERT(console_resource->password_.encoding == p_encoding_md5);
-    password = &console_resource->password_;
     local_tls_resource = console_resource;
   } else { /* default console */
-    name = "*UserAgent*";
+    name = DEFAULT_CONSOLE_NAME;
     ASSERT(director_resource->password_.encoding == p_encoding_md5);
-    password = &director_resource->password_;
     local_tls_resource = director_resource;
   }
 
-  if (local_tls_resource->IsTlsConfigured()) {
-    std::string qualified_resource_name = global_resource::QualifiedName(
-        my_config->GlobalTypeFromLocalType(my_config->r_own_), name);
-    if (qualified_resource_name.empty()) {
-      delete UA_sock;
-      UA_sock = nullptr;
-      jcr.dir_bsock = nullptr;
-      return nullptr;
-    }
-
-    if (!UA_sock->DoTlsHandshake(TlsPolicy::kBnetTlsAuto, local_tls_resource,
-                                 false, qualified_resource_name.c_str(),
-                                 password->value, &jcr)) {
-      delete UA_sock;
-      UA_sock = nullptr;
-      jcr.dir_bsock = nullptr;
-      return nullptr;
-    }
-  } /* IsTlsConfigured */
-
-  std::string own_qualified_name = "R_CONSOLE::";
-  own_qualified_name += name;
-
-  if (!UA_sock->ConsoleAuthenticateWithDirector(
-          &jcr, name, *password, director_resource, own_qualified_name,
-          response_args, response_id)) {
+  if (!BareosConnect<global_resource::Type::Console,
+                     global_resource::Type::Director>(&jcr, UA_sock, name,
+                                                      local_tls_resource)) {
     delete UA_sock;
     UA_sock = nullptr;
     jcr.dir_bsock = nullptr;
     return nullptr;
   }
+
+  if (!UA_sock->ReceiveAndEvaluateResponseMessage(response_id, response_args)) {
+    delete UA_sock;
+    UA_sock = nullptr;
+    jcr.dir_bsock = nullptr;
+    return nullptr;
+  }
+
   return UA_sock;
 }