+++ core/src/dird/ua_server.h
@@ -1,7 +1,7 @@
 /*
    BAREOS® - Backup Archiving REcovery Open Sourced
 
-   Copyright (C) 2018-2020 Bareos GmbH & Co. KG
+   Copyright (C) 2018-2026 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -26,8 +26,9 @@ class BareosSocket;
 class JobControlRecord;
 
 namespace directordaemon {
+class UaContext;
 
-void* HandleUserAgentClientRequest(BareosSocket* user);
+void* HandleUserAgentClientRequest(UaContext* ua);
 JobControlRecord* new_control_jcr(const char* base_name, int job_type);
 
 } /* namespace directordaemon */