+++ core/src/stored/authenticate.cc
@@ -26,9 +26,12 @@
  */
 
 #include "include/bareos.h"
+#include "lib/s_password.h"
 #include "stored/stored.h"
 #include "stored/stored_globals.h"
 #include "stored/stored_jcr_impl.h"
+#include "stored/authenticate.h"
+#include <climits>
 #include "lib/parse_conf.h"
 #include "lib/bsock.h"
 #include "lib/bnet_network_dump.h"
@@ -37,194 +40,85 @@
 
 namespace storagedaemon {
 
-const int debuglevel = 50;
-
-constexpr const char Dir_sorry[] = "3999 No go\n";
-constexpr const char OK_hello[] = "3000 OK Hello\n";
-
-/**
- * Initiate the message channel with the Director.
- * It has made a connection to our server.
- *
- * Basic tasks done here:
- *   - Assume the Hello message is already in the input buffer.
- *   - Authenticate
- *   - Get device
- *   - Get media
- *   - Get pool information
- *
- * This is the channel across which we will send error
- * messages and job status information.
- */
-bool AuthenticateDirector(JobControlRecord* jcr)
+const TlsResource* Auth::get(global_resource::Type auth_type,
+                             std::string_view name)
 {
-  BareosSocket* dir = jcr->dir_bsock;
-  POOLMEM* dirname;
-  DirectorResource* director = NULL;
+  auto* myself
+      = dynamic_cast<StorageResource*>(p->GetNextRes(R_STORAGE, nullptr));
 
-  // Sanity check.
-  if (dir->message_length < 25 || dir->message_length > 500) {
-    Dmsg2(debuglevel, "Bad Hello command from Director at %s. Len=%d.\n",
-          dir->who(), dir->message_length);
-    Jmsg2(jcr, M_FATAL, 0,
-          T_("Bad Hello command from Director at %s. Len=%d.\n"), dir->who(),
-          dir->message_length);
-    return false;
-  }
-  dirname = GetPoolMemory(PM_MESSAGE);
-  dirname = CheckPoolMemorySize(dirname, dir->message_length);
-
-  unsigned major = 0, minor = 0, patch = 0;
-
-  if (bsscanf(dir->msg, "Hello Director %127s calling Version=\"%u.%u.%u\"",
-              dirname, &major, &minor, &patch)
-      == 4) {
-    dir->remote_version = VERSION_HEX(major, minor, patch);
-  } else if (bsscanf(dir->msg, "Hello Director %127s calling", dirname) == 1) {
-  } else {
-    dir->msg[100] = 0;
-    Dmsg2(debuglevel, "Bad Hello command from Director at %s: %s\n", dir->who(),
-          dir->msg);
-    Jmsg2(jcr, M_FATAL, 0, T_("Bad Hello command from Director at %s: %s\n"),
-          dir->who(), dir->msg);
-    FreePoolMemory(dirname);
-    return false;
+  if (!myself) {
+    Emsg1(M_ERROR, 0, "Could not find myself during connection attempt.\n");
+    return nullptr;
   }
 
-  UnbashSpaces(dirname);
-  director = (DirectorResource*)my_config->GetResWithName(R_DIRECTOR, dirname);
-  jcr->sd_impl->director = director;
+  int name_len = (int)std::min((size_t)INT_MAX, name.size());
+  const char* name_ptr = name.data();
+  char tbuf[MAX_TIME_LENGTH];
 
-  if (!director) {
-    Dmsg2(debuglevel, "Connection from unknown Director %s at %s rejected.\n",
-          dirname, dir->who());
-    Jmsg(jcr, M_FATAL, 0,
-         T_("Connection from unknown Director %s at %s rejected.\n"), dirname,
-         dir->who());
-    FreePoolMemory(dirname);
-    return false;
+  switch (auth_type) {
+    case global_resource::Type::Job: {
+      auto* jcr = get_jcr_for_authentication(name);
+      if (!jcr) {
+        Dmsg1(50, "Authentication for job %.*s not possible\n", name_len,
+              name_ptr);
+        Emsg1(M_ERROR, 0, T_("Connection for unknown job %.*s rejected.\n"),
+              name_len, name_ptr);
+        return nullptr;
+      }
+
+      auto& data = job.emplace();
+      data.jcr = jcr;
+      data.job = *myself;
+      data.job.password_.value = jcr->sd_auth_key;
+
+      type = inbound_type::Job;
+      return &data.job;
+    } break;
+    case global_resource::Type::Director: {
+      auto* res = dynamic_cast<DirectorResource*>(
+          p->GetResWithName(R_DIRECTOR, name));
+      if (!res) {
+        Dmsg1(60, "Got a DIR connection at %s from unknown dir %.*s\n",
+              bstrftimes(tbuf, sizeof(tbuf), (utime_t)time(NULL)), name_len,
+              name_ptr);
+        Emsg1(M_ERROR, 0,
+              T_("Connection from unknown Director %.*s rejected.\n"), name_len,
+              name_ptr);
+        return nullptr;
+      }
+
+      if (res->password_.encoding != p_encoding_md5) {
+        Emsg1(M_ERROR, 0, "Bad password type for dir %.*s\n", name_len,
+              name_ptr);
+        return nullptr;
+      }
+
+      {
+        JobControlRecord* ojcr;
+        unsigned int cnt = 0;
+
+        foreach_jcr (ojcr) { cnt++; }
+        endeach_jcr(ojcr);
+
+        if (cnt >= myself->MaxConcurrentJobs) {
+          Emsg0(M_ERROR, 0,
+                T_("Number of Jobs exhausted, please increase "
+                   "MaximumConcurrentJobs\n"));
+          return nullptr;
+        }
+      }
+
+
+      auto& data = director.emplace();
+      data.res = res;
+
+      type = inbound_type::Director;
+      return res;
+    } break;
+    default: {
+    } break;
   }
 
-  if (!dir->AuthenticateInboundConnection(jcr, my_config,
-                                          director->resource_name_,
-                                          director->password_, director)) {
-    dir->fsend("%s", Dir_sorry);
-    Dmsg2(debuglevel, "Unable to authenticate Director \"%s\" at %s.\n",
-          director->resource_name_, dir->who());
-    Jmsg1(jcr, M_ERROR, 0, T_("Unable to authenticate Director at %s.\n"),
-          dir->who());
-    Bmicrosleep(5, 0);
-    FreePoolMemory(dirname);
-    return false;
-  }
-
-  FreePoolMemory(dirname);
-  return dir->fsend("%s", OK_hello);
+  return nullptr;
 }
-
-/**
- * Authenticate a remote storage daemon.
- *
- * This is used for SD-SD replication of data.
- */
-bool AuthenticateStoragedaemon(JobControlRecord* jcr)
-{
-  BareosSocket* sd = jcr->store_bsock;
-  s_password password;
-
-  password.encoding = p_encoding_md5;
-  password.value = jcr->sd_auth_key;
-  const char* identity = "* replicate *";
-
-  Dmsg2(debuglevel, "AuthenticateStoragedaemon %s %s\n", identity,
-        (unsigned char*)password.value);
-  if (!sd->AuthenticateInboundConnection(jcr, my_config, identity, password,
-                                         me)) {
-    Jmsg1(jcr, M_FATAL, 0,
-          T_("Authorization problem: Two way security handshake failed with "
-             "Storage daemon at %s\n"),
-          sd->who());
-    return false;
-  }
-
-  return true;
-}
-
-/**
- * Authenticate with a remote storage daemon.
- *
- * This is used for SD-SD replication of data.
- */
-bool AuthenticateWithStoragedaemon(JobControlRecord* jcr)
-{
-  BareosSocket* sd = jcr->store_bsock;
-  s_password password;
-
-  const char* identity = "* replicate *";
-  password.encoding = p_encoding_md5;
-  password.value = jcr->sd_auth_key;
-
-  if (!sd->AuthenticateOutboundConnection(
-          jcr, my_config->CreateOwnQualifiedNameForNetworkDump(), identity,
-          password, me)) {
-    Jmsg1(jcr, M_FATAL, 0,
-          T_("Authorization problem: Two way security handshake failed with "
-             "Storage daemon at %s\n"),
-          sd->who());
-    return false;
-  }
-
-  return true;
-}
-
-/**
- * Authenticate a remote File daemon.
- *
- * This is used for FD backups or restores.
- */
-bool AuthenticateFiledaemon(JobControlRecord* jcr)
-{
-  BareosSocket* fd = jcr->file_bsock;
-  s_password password;
-
-  password.encoding = p_encoding_md5;
-  password.value = jcr->sd_auth_key;
-
-  if (!fd->AuthenticateInboundConnection(jcr, my_config, jcr->client_name,
-                                         password, me)) {
-    Jmsg1(jcr, M_FATAL, 0,
-          T_("Authorization problem: Two way security handshake failed with "
-             "File daemon at %s\n"),
-          fd->who());
-    return false;
-  }
-
-  return true;
-}
-
-/**
- * Authenticate with a remote file daemon.
- *
- * This is used for passive FD backups or restores.
- */
-bool AuthenticateWithFiledaemon(JobControlRecord* jcr)
-{
-  BareosSocket* fd = jcr->file_bsock;
-  s_password password;
-
-  password.encoding = p_encoding_md5;
-  password.value = jcr->sd_auth_key;
-
-  if (!fd->AuthenticateOutboundConnection(jcr, "File daemon", jcr->client_name,
-                                          password, me)) {
-    Jmsg1(jcr, M_FATAL, 0,
-          T_("Authorization problem: Two way security handshake failed with "
-             "File daemon at %s\n"),
-          fd->who());
-    return false;
-  }
-
-  return true;
-}
-
 } /* namespace storagedaemon */