+++ core/src/stored/dir_cmd.h
@@ -1,7 +1,7 @@
 /*
    BAREOS® - Backup Archiving REcovery Open Sourced
 
-   Copyright (C) 2018-2018 Bareos GmbH & Co. KG
+   Copyright (C) 2018-2026 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -21,9 +21,11 @@
 #ifndef BAREOS_STORED_DIR_CMD_H_
 #define BAREOS_STORED_DIR_CMD_H_
 
+#include "lib/bsock.h"
+#include "stored/stored_conf.h"
 namespace storagedaemon {
 
-void* HandleDirectorConnection(BareosSocket* dir);
+void* HandleDirectorConnection(BareosSocket* dir, DirectorResource* res);
 
 } /* namespace storagedaemon */